Uploaded ClosestPrinter and PrinterUtil modules
A note on selecting a map/image for the campus:

Alec and I noticed that a number of campus maps that show up in Google
image search are not in the proper alignment and have been rotated as
much as 45 degrees sometimes.  Easiest/safest thing to do would be to
screenshot the Google Maps map view.  Otherwise, just verify that any
image you choose has proper North, South, East, West alignment